+++ teacher.rb
@@ -1,0 +1,12 @@
+require_relative 'person'
+
+class Teacher < Person
+  def initialize(specialization, name = 'unknown', age = nil, parent_permission: true)
+    super(name, age, parent_permission: parent_permission)
+    @specialization = specialization
+  end
+
+  def can_use_services?
+    true
+  end
+end
